Add test helpers and fix up test stubs
