+++ test/unit/handlers/api/handler.test.js
@@ -3,54 +3,19 @@
 const Test = require('tapes')(require('tape'))
 const Sinon = require('sinon')
 const request = require('request-promise-native')
-const HealthCheck = require('@mojaloop/central-services-shared').HealthCheck.HealthCheck
 
 const Notification = require('../../../../src/handlers/notification')
 const Handler = require('../../../../src/api/metadata/handler')
-
-function createRequest (routes) {
-  let value = routes || []
-  return {
-    server: {
-      table: () => {
-        return [{ table: value }]
-      }
-    }
-  }
-}
-
-/**
- * unwrapResponse
- *
- * Use this function to unwrap the innner response body and code from an async Handler
- */
-const unwrapResponse = async (asyncFunction) => {
-  let responseBody
-  let responseCode
-  const nestedReply = {
-    response: (response) => {
-      responseBody = response
-      return {
-        code: statusCode => {
-          responseCode = statusCode
-        }
-      }
-    }
-  }
-  await asyncFunction(nestedReply)
-
-  return {
-    responseBody,
-    responseCode
-  }
-}
+const {
+  createRequest,
+  unwrapResponse
+} = require('../../../helpers')
 
 Test('route handler', (handlerTest) => {
   let sandbox
 
   handlerTest.beforeEach(t => {
     sandbox = Sinon.createSandbox()
-    sandbox.stub(HealthCheck.prototype, 'getHealth').resolves()
     sandbox.stub(Notification, 'isConnected')
     sandbox.stub(request, 'get')
 
@@ -66,7 +31,6 @@ Test('route handler', (handlerTest) => {
   handlerTest.test('/health should', healthTest => {
     healthTest.test('returns the correct response when the health check is up', async test => {
       // Arrange
-      // HealthCheck.prototype.getHealth.resolves({ status: 'OK' })
       Notification.isConnected.resolves(true)
       request.get.resolves({status: 'OK'})
       const expectedResponseCode = 200
@@ -83,8 +47,8 @@ Test('route handler', (handlerTest) => {
 
     healthTest.test('returns the correct response when the health check is down', async test => {
       // Arrange
-      // HealthCheck.prototype.getHealth.throws(new Error('getHealth() failed'))
       Notification.isConnected.throws(new Error('Error connecting to consumer'))
+      request.get.resolves({ status: 'OK' })
 
       const expectedResponseCode = 502
